feat(front): Résolution problème query comment
