+++ src/components/AllPosts.vue
@@ -3,8 +3,8 @@
     <v-row justify="center">
       <v-col cols="12" sm="6" class="pt-0">
         <v-card
-          v-for="(post, index) of posts"
-          :key="index"
+          v-for="post of posts"
+          :key="post.id"
           outlined
           elevation="2"
           id="post--card"
@@ -95,7 +95,56 @@
             </div>
           </div>
 
-          <NewComment />
+          <div>
+            <v-form
+              @submit.prevent="publishComment"
+              class="d-flex flex-column justify-center pT-2 pr-2"
+            >
+              <div id="comment--wrapper" class="mt-2 mb-2 mx-4">
+                <v-textarea
+                  v-model="text"
+                  name="text"
+                  rounded
+                  single-line
+                  auto-grow
+                  rows="1"
+                  hide-details="auto"
+                  label="Ajouter un commentaire..."
+                  color="deep-purple"
+                  class="ma-0 pt-1"
+                >
+                </v-textarea>
+              </div>
+
+              <div class="d-flex">
+                <v-file-input
+                  @change="uploadFile"
+                  type="file"
+                  name="image"
+                  ref="imageInput"
+                  accept="image/png, image/jpeg, image/bmp, image/webp, image/gif"
+                  prepend-icon="mdi-image"
+                  color="deep-purple"
+                  class="mx-4 py-0"
+                  small-chips
+                  required
+                ></v-file-input>
+
+                <v-btn
+                  type="submit"
+                  :id="post.id"
+                  :value="post.id"
+                  elevation="2"
+                  color="deep-purple"
+                  class="white--text"
+                  rounded
+                  >Publier</v-btn
+                >
+              </div>
+            </v-form>
+
+            <v-divider class="mx-4"></v-divider>
+          </div>
 
           <div class="d-flex ma-4">
             <v-avatar color="deep-purple">
@@ -127,15 +176,17 @@
 </template>
 
 <script>
-import NewComment from "./NewComment.vue"
-
 import axios from "axios"
 
 export default {
-  components: { NewComment },
   data() {
     return {
       posts: [],
+      text: "",
+      idUser: "",
+      idPost: "",
+      username: "",
+      image: null,
     }
   },
 
@@ -148,6 +199,36 @@ export default {
 
   methods: {
     deletePost() {},
+    uploadFile() {
+      this.image = this.$refs.imageInput[0].$refs.input.files[0]
+    },
+    publishComment() {
+      let formData = new FormData()
+
+      let user = JSON.parse(localStorage.getItem("user"))
+      this.idUser = user.id
+      this.username = user.username
+
+      this.idPost = this.posts[0].id
+
+      formData.append("idPost", this.idPost)
+      formData.append("idUser", this.idUser)
+      formData.append("username", this.username)
+      formData.append("text", this.text)
+      formData.append("image", this.image)
+
+      const datata = [...formData.entries()]
+      console.log(datata)
+
+      axios
+        .post("//localhost:3000/api/comment/createcomment", formData)
+        .then(() => {
+          console.log("Post créé.")
+        })
+        .catch((error) => console.log(error))
+
+      // window.location.reload()
+    },
   },
 }
 </script>
@@ -184,5 +265,16 @@ p {
 #post--content {
   font-size: 1rem;
   color: black;
+}
+
+#comment {
+  &--wrapper {
+    border: 1px solid grey;
+    border-radius: 28px;
+  }
+  &--date {
+    line-height: 1rem;
+    font-size: 1rem;
+  }
 }
 </style>
